Remove POC Module (#2945)
