+++ sdv/utils/__init__.py
@@ -3,13 +3,17 @@
 from sdv.utils.utils import (
     drop_unknown_references,
     get_random_sequence_subset,
-    load_synthesizer,
+    get_random_subset,
     load_constraints,
+    load_synthesizer,
+    simplify_schema,
 )
 
 __all__ = (
     'drop_unknown_references',
     'get_random_sequence_subset',
+    'get_random_subset',
     'load_synthesizer',
     'load_constraints',
+    'simplify_schema',
 )
